UPSTREAM: bpf: revert report offload info to user space
This reverts commit bd601b6ada11 ("bpf: report offload info to user
space").  The ifindex by itself is not sufficient, we should provide
information on which network namespace this ifindex belongs to.
After considering some options we concluded that it's best to just
remove this API for now, and rework it in -next.

Change-Id: Id934c8a26171e4477ccbdf20ee38287db6b73f45
Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Naveen <133593113+elohim-etz@users.noreply.github.com>
(cherry picked from commit afac6fab8383a4bbc3e3642a44511a113fbb6d2e)