UPSTREAM: stddef.h: Introduce sizeof_field()
The size of fields within a structure is needed in a few places in the
kernel already, and will be needed for the usercopy whitelisting when
declaring whitelist regions within structures. This creates a dedicated
macro and redefines offsetofend() to use it.

Existing usage, ignoring the 1200+ lustre assert uses:

$ git grep -E 'sizeof\(\(\((struct )?[a-zA-Z_]+ \*\)0\)->' | \
	grep -v staging/lustre | wc -l
65

Change-Id: I68f46fefe29fb6c9dfe358717a343e108f797d81
Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Naveen <133593113+elohim-etz@users.noreply.github.com>
(cherry picked from commit 77a92d218216c5d0ed1eb899392e401689ca17d3)